Move transmission to a separate file
